% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diamond.R
\name{read_diamond}
\alias{read_diamond}
\title{Import DIAMOND2 output file}
\usage{
read_diamond(file, out_format, use_arrow_duckdb_connection = FALSE)
}
\arguments{
\item{file}{path to BLAST output file.}

\item{out_format}{a character string specifying the output format of the BLAST output that shall be imported.
Available options are:
 \itemize{
 \item \code{out_format = "xml"} : XML
 \item \code{out_format = "csv"} : Comma-separated values
 }}

\item{use_arrow_duckdb_connection}{shall DIAMOND2 hit output table be transformed to an in-process (big data disk-processing) arrow connection to DuckDB? This is useful when the DIAMOND2 output table to too large to fit into memory. Default is \code{use_arrow_duckdb_connection = FALSE}.}
}
\description{
When performing BLAST searches with the \code{blast_*()} functions,
the corresponding BLAST output file can be imported into the current R session using this function.
All output formats given by BLAST are supported (see e.g. description for details).
}
\seealso{
\code{\link{diamond_protein_to_protein}}
}
\author{
Hajk-Georg Drost
}
